Add tinymap and arraymap
