+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tinymap"
-version = "0.2.4"
+version = "0.3.0"
 authors = ["not_a_seagull <jtnunley01@gmail.com>"]
 edition = "2018"
 description = "A map structure that stores its data on the stack"
@@ -9,4 +9,9 @@ repository = "https://github.com/not-a-seagull/tinymap"
 homepage = "https://github.com/not-a-seagull/tinymap#readme"
 
 [dependencies]
+hashbrown = { version = "0.8.2", optional = true }
 tinyvec = { version = "1.0.0", features = ["nightly_const_generics"] }
+
+[features]
+default = []
+alloc = ["hashbrown"]
